fix tidy

+++ tasks/frolova_s_star_topology/seq/src/ops_seq.cpp
@@ -30,7 +30,7 @@ bool FrolovaSStarTopologySEQ::RunImpl() {
     tmp.push_back(i);
   }
 
-  GetOutput() = tmp.size();
+  GetOutput() = static_cast<int>(tmp.size());
 
   return true;
 }
